requirements.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.2.1
pymongo==3.12.3
pytz==2022.6
snappy==2.4.1
sqlparse==0.2.4
tzdata==2022.6
